+++ Changelog.md
@@ -14,6 +14,7 @@
 - Add expression: getinventorycountbytemplatetype[double templateType]
 - Add support for wands with built in spells with actiontryapplyitem and ub use (thanks aquafir)
 - Add [m] flag to ub mexec command for disabling chat output (thanks aquafir)
+- Add expressions: uboptset[name, newValue], uboptget[name]
 
 ## 0.2.4 (2021-03-29) [UtilityBeltInstaller-0.2.4.exe](https://gitlab.com/utilitybelt/utilitybelt.gitlab.io/uploads/deac20cf3c483c039174e4e9fee2f7ca/UtilityBeltInstaller-0.2.4.exe)
 - Fix classic server vtank patch, no longer patches non-itemenchantment auras